Fixed compiler warning and added new features.

+++ Pod/Classes/NSArray+BlockExtensions.m
@@ -13,7 +13,7 @@
 
 -(id)randomObject {
     if (self.count == 0) return nil;
-    return [self objectAtIndex:arc4random_uniform(self.count)];
+    return [self objectAtIndex:arc4random_uniform((u_int32_t)self.count)];
 }
 
 @end
